- fixed google search size from int to long

+++ app/src/main/java/com/bnnvara/kiespijn/GoogleImageSearch/GoogleSearchFragment.java
@@ -48,7 +48,7 @@ public class GoogleSearchFragment extends Fragment {
     private List<ImageView> mImageViewList;
     private android.widget.SearchView mSearchView;
     private int mImageIndex = 1;
-    private int mTotalImageSize;
+    private long mTotalImageSize;
     private MenuItem mNewSearch;
 
     public static GoogleSearchFragment newInstance(String searchString) {
@@ -67,7 +67,6 @@ public class GoogleSearchFragment extends Fragment {
         if (mSearchString != null) {
             getImages();
         }
-
 
     }
